Move shared methods to tile-loader-utils.ts
